pyproject.toml
[tool.poetry]
name = "poetry-lab"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[[tool.poetry.source]]
name = "pt-security"
url = "https://pypi-proxy.ptsecurity.com"
priority = "default"

[tool.poetry.dependencies]
python = "^3.10"
allure-pytest = "^2.13.2"
assertpy = "^1.1"
pytest = "^7.4.0"
pytest-playwright = "^0.3.3"
pytest-xdist = "^3.3.1"
python-dotenv = "^1.0.0"
pydantic = "^1.10.11"

[tool.poetry.group.lint.dependencies]
mypy = "1.3.0"
flake8-pyproject = "^1.2.3"
flake8-black = "^0.3.6"
flake8-isort = "^6.0.0"

[tool.poetry.group.lint]
optional = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s]\nMESSAGE: %(message)s\nLOCATON: (%(filename)s:%(lineno)s)\n"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "--strict-markers"
markers = [
    "smoke",
]

[tool.isort]
profile = "black"

[tool.flake8]
ignore = ['E203']
max-line-length = 120

[tool.mypy]
# 3rd party import
ignore_missing_imports = true
# dynamic typing
disallow_any_unimported = true
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = true
disallow_any_generics = false
disallow_subclassing_any = true
# platform
python_version = "3.10"
# untyped
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
# None and Optional
no_implicit_optional = true
# Warnings
warn_return_any = false
warn_unreachable = true
# Misc
pretty = true
